CascadeProjects/windsurf-project/requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
deepface==0.0.99
dill==0.4.1
distro==1.9.0
easydict==1.13
fastapi==0.136.1
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.7.0
gdown==5.2.1
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-genai==1.68.0
google-pasta==0.2.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
gunicorn==25.1.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.9.2
ibm-cloud-sdk-core==3.24.4
ibm-platform-services==0.74.0
ibm-quantum-schemas==0.5.20260320
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
keras==3.13.2
kiwisolver==1.5.0
langchain==1.2.13
langchain-core==1.2.21
langchain-openai==1.1.12
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lazy-loader==0.5
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.1
namex==0.1.0
networkx==3.6.1
numpy==2.4.3
oauthlib==3.3.1
onnx==1.20.1
onnxruntime==1.24.4
openai==2.29.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
peft==0.18.1
pillow==12.1.1
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
PySocks==1.7.1
pyspnego==0.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
qiskit==2.3.1
qiskit-ibm-runtime==0.46.1
regex==2026.2.28
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
retina-face==0.0.17
rich==14.3.3
rustworkx==0.17.1
safetensors==0.7.0
samplomatic==0.17.1
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8.3
starlette==1.0.0
stevedore==5.7.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
tf_keras==2.21.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
twilio==9.10.5
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
